+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### ✅ Added
 - Possibility to customize message reactions top padding (for grid-based reaction containers)
 - Custom sorting of reactions
+- Added a configurable separator view for new messages
 
 # [4.26.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.26.0)
 _January 16, 2023_